@@ -8,11 +8,11 @@
     "homepage": "https://github.com/eighty8/Stringy",
     "license": "MIT",
     "require": {
-        "php": ">=5.4.0",
+        "php": ">=7.2",
         "symfony/polyfill-mbstring": "~1.1"
     },
     "require-dev": {
-        "phpunit/phpunit": "~4.0"
+        "phpunit/phpunit": "~8.0"
     },
     "support": {
         "issues": "https://github.com/eighty8/Stringy/issues",
