@@ -2,13 +2,12 @@
 
 use Stringy\StaticStringy as S;
 
-class StaticStringyTestCase extends PHPUnit_Framework_TestCase
+class StaticStringyTestCase extends \PHPUnit\Framework\TestCase
 {
-    /**
-     * @expectedException BadMethodCallException
-     */
+
     public function testBadMethodCall()
     {
+        $this->expectException(BadMethodCallException::class);
         $result = S::invalidMethod('foo');
     }
 
@@ -22,30 +21,30 @@ class StaticStringyTestCase extends PHPUnit_Framework_TestCase
     {
         $result = S::toLowerCase('FOOBAR');
         $this->assertEquals('foobar', $result);
-        $this->assertInternalType('string', $result);
+        $this->assertIsString($result);
     }
 
     public function testPartialArgsInvocation()
     {
         $result = S::slice('foobar', 0, 3);
         $this->assertEquals('foo', $result);
-        $this->assertInternalType('string', $result);
+        $this->assertIsString($result);
     }
 
     public function testFullArgsInvocation()
     {
         $result = S::slice('fòôbàř', 0, 3, 'UTF-8');
         $this->assertEquals('fòô', $result);
-        $this->assertInternalType('string', $result);
+        $this->assertIsString($result);
     }
 
     public function testArrayReturnValue()
     {
         $result = S::lines("a\nb");
         $this->assertEquals(['a', 'b'], $result);
-        $this->assertInternalType('array', $result);
+        $this->assertIsArray($result);
         foreach ($result as $val) {
-            $this->assertInternalType('string', $val);
+            $this->assertIsString($val);
         }
     }
 
@@ -64,7 +63,7 @@ class StaticStringyTestCase extends PHPUnit_Framework_TestCase
 
         foreach ($properties['methodArgs'] as $method => $expected) {
             $num = $stringyClass->getMethod($method)
-                ->getNumberOfParameters() + 2;
+                    ->getNumberOfParameters() + 2;
 
             $this->assertEquals($expected, $num,
                 'Invalid num args for ' . $method);
